refactored Container component
