--- src/common/Container/index.js
@@ -1,10 +1,0 @@
-import React from "react";
-import { ContainerMain } from "./styled";
-
-const Container = ({ children }) => (
-  <ContainerMain>
-    {children}
-  </ContainerMain>
-)
-
-export default Container;
