+++ src/common/Container/styled.js
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 
-export const ContainerMain = styled.main`
+export const Container = styled.main`
   max-width: 900px;
   margin: 0 auto;
   padding: 20px;
