+++ src/features/tasks/TasksPage/index.js
@@ -4,9 +4,9 @@ import Form from "./Form";
 import Buttons from "./Buttons";
 import TaskList from "./TaskList";
 import Section from "../../../common/Section";
-import Container from "../../../common/Container";
 import ExampleTasksButton from "./ExampleTasksButton";
 import Search from "./Search";
+import { Container } from '../../../common/Container/styled';
 
 const Tasks = () => (
   <Container>
